[DEVX-1176] FxPluginRequest -> FxPlugin::Request
