+++ lib/global_reach_partners/fx_plugin.rb
@@ -1,0 +1,1 @@
+require 'global_reach_partners/fx_plugin/request'
