+++ lib/global_reach_partners/fx_plugin/request.rb
@@ -1,0 +1,18 @@
+module GlobalReachPartners
+  module FxPlugin
+    class Request < GlobalReachPartners::Request
+      def client
+        GlobalReachPartners.fx_plugin_client
+      end
+
+      def authentication
+        {
+          client_code: configuration.client_code,
+          username: configuration.username,
+          password: configuration.password,
+          group_ID: GRP
+        }
+      end
+    end
+  end
+end
