--- lib/global_reach_partners/fx_plugin_request.rb
@@ -1,16 +1,0 @@
-module GlobalReachPartners
-  class FxPluginRequest < Request
-    def client
-      GlobalReachPartners.fx_plugin_client
-    end
-
-    def authentication
-      {
-        client_code: configuration.client_code,
-        username: configuration.username,
-        password: configuration.password,
-        group_ID: GRP
-      }
-    end
-  end
-end
